requirements.txt
fastapi[standard]
langchain_cloudflare==0.1.8
langchain_core==0.3.60
langchain_groq==0.3.2
langchain_qdrant==0.2.0
passlib==1.7.4
Pillow==11.2.1
pydantic==2.11.4
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.13.0
pymupdf==1.25.5
pytesseract==0.3.13
python-dotenv==1.1.0
qdrant_client==1.14.2
supabase==2.15.1
